% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{black_hills_pop}
\alias{black_hills_pop}
\title{Spatially indexed forest nonforest probabilities for the Black Hills in South Dakota and Wyoming.}
\format{A data frame with 24240 rows and 3 variables:
\describe{
  \item{forest_nonforest_prob}{probability of forest}
  \item{s1}{spatial grid index 1}
  \item{s2}{spatial grid index 2}
}}
\source{
\url{https://data.fs.usda.gov/geodata/rastergateway/}
}
\usage{
black_hills_pop
}
\description{
A dataset containing forest nonforest probabilities across the Black Hills from geospatial satellite raster data.
}
\keyword{datasets}
