% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{black_hills_samp}
\alias{black_hills_samp}
\title{Spatially indexed biomass and forest nonforest probability samples from the Black Hills in South Dakota and Wyoming.}
\format{A data frame with 232 rows and 4 variables:
\describe{
  \item{forest_nonforest_prob}{probability of forest}
  \item{biomass}{forest biomass in Megagrams per Hectare}
  \item{s1}{spatial grid index 1}
  \item{s2}{spatial grid index 2}
}}
\source{
\url{https://data.fs.usda.gov/geodata/rastergateway/}
}
\usage{
black_hills_samp
}
\description{
A sample of biomass measures and forest nonforest probabilities from geospatial satellite raster data.
}
\keyword{datasets}
