% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregEstimate.R
\name{gregEstimate}
\alias{gregEstimate}
\title{Compute a generalized regression estimator}
\usage{
gregEstimate(y, y_hat_sample, pi = NULL, y_hat_pop, pi2 = NULL,
  N = NULL, data_type = "raw", var_est = FALSE,
  var_method = "lin_HB", B = 1000, strata = NULL)
}
\arguments{
\item{y}{A numeric vector of the sampled response variable.}

\item{y_hat_sample}{A vector of predicted sample estimates from a survey model.}

\item{pi}{A numeric vector of inclusion probabilities for each sampled unit in y.  If NULL, then simple random sampling without replacement is assumed.}

\item{y_hat_pop}{If data_type is "raw", y_hat_pop must be a vector of predicted population estimates from a survey model. Otherwise it must be either the total or mean population estimate.}

\item{pi2}{A square matrix of the joint inclusion probabilities.  Needed for the "lin_HT" variance estimator.}

\item{N}{A numeric value of the population size. If NULL, it is estimated with the sum of the inverse of the pis.}

\item{data_type}{A string that specifies the form of population auxiliary data. The possible values are "raw", "totals" or "means" for whether the user is providing population estimates at the unit level, aggregated to totals, or aggregated to means.  Default is "raw".}

\item{var_est}{A logical indicating whether or not to compute a variance estimator.  Default is FALSE.}

\item{var_method}{The method to use when computing the variance estimator.  Options are a Taylor linearized technique: "lin_HB"= Hajek-Berger estimator, "lin_HH" = Hansen-Hurwitz estimator, "lin_HTSRS" = Horvitz-Thompson estimator under simple random sampling without replacement, and "lin_HT" = Horvitz-Thompson estimator or a resampling technique: "bootstrap_SRS" = bootstrap variance estimator under simple random sampling without replacement. The default is "lin_HB".}

\item{B}{The number of bootstrap samples if computing the bootstrap variance estimator.  Default is 1000.}

\item{strata}{A factor vector of the stratum membership.  If NULL, all units are put into the same stratum.  Must have same length as y.}
}
\value{
A greg object containing:
\itemize{
\item{pop_total: Estimate of population total}
\item{pop_mean: Estimate of the population mean}
\item{pop_total_var: Estimated variance of population total estimate}
\item{pop_mean_var: Estimated variance of population mean estimate}
\item{weights: Survey weights produced by greg (linear model only)}
\item{coefficients: Survey-weighted model coefficients}
}
}
\description{
Summarizes population estimates given precomputed predicted sample and population values.
}
