% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregObject.R
\name{plot.greg}
\alias{plot.greg}
\title{Compute a geospatial heat map to summarize a generalized regression estimator}
\usage{
\method{plot}{greg}(obj, x_sample, x_pop, spatial_1, spatial_2,
  bins = 300, fun = function(x) mean(x), hexagon = FALSE)
}
\arguments{
\item{obj}{Object of class 'greg'.}

\item{x_sample}{A data frame of the sample level data with two spatial variables. Must have the same indicies as the original data frame used to train model.}

\item{x_pop}{A data frame of the population level data with two spatial variables. Must have the same indicies as the original population data frame used to make model.}

\item{spatial_1}{A string for the name of variable in `x_sample` and `x_pop` to plot as x axis (column name of longitude).}

\item{spatial_2}{A string for the name of variable in `x_sample` and `x_pop` to plot as y axis (column name of latitude).}

\item{bins}{An integer specifying the number of bins to aggregate estimated values}

\item{fun}{Summary function to apply within each bin to be plotted. Default is `mean(x)`. See `stat_summary`.}

\item{hexagon}{Boolean. Specify if you want hexagonal shaped bins or not.}
}
\description{
Computes a geospatial heat map of fitted values of a 'greg' object
}
