% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postStrat.R
\name{postStrat}
\alias{postStrat}
\title{Compute a post-stratified estimator}
\usage{
postStrat(y, xsample, xpop, pi = NULL, N = NULL, var_est = FALSE,
  var_method = "HB", pi2 = NULL, datatype = "raw", B = 1000)
}
\arguments{
\item{y}{vector or a matrix with one column of the sampled response variable}

\item{xsample}{vector or a matrix with one column of an auxiliary categorical variable giving the stratum of each sampled unit.}

\item{xpop}{vector or a matrix with one column of an auxiliary categorical variable giving the stratum of each population unit.}

\item{pi}{Default to assume equal probability/simple random sampling, if unequal probability, requires vector of first-order inclusion probabilities of same length as y}

\item{N}{population size, if not provided estimated to be the sum of the inverse inclusion probabilities}

\item{var_est}{Default to FALSE, logical for whether or not to compute estimate of variance}

\item{var_method}{Method to use when computing the variance estimate.  Options are "HB"= Hajek-Berger estimator, "HH" = Hansen-Hurwitz estimator, "HTSRS" = Horvitz-Thompson estimator under simple random sampling, "HT" = Horvitz-Thompson estimator, "bootstrapSRS" = bootstrap variance estimator under simple random sampling without replacement, "SRSunconditional" = simple random sampling variance estimator which accounts for random strata}

\item{pi2}{a square matrix of the joint inclusion probabilities.  Needed for the "HT" variance estimator}

\item{datatype}{Default to "raw", takes values "raw", "totals" or "means" for whether the user is providing the raw population stratum memberships, the population totals of each stratum, or the population proportions of each stratum (MAKE SURE TO DISCUSS FORM var_cat1, var_car2)}

\item{B}{number of bootstrap samples if computing the bootstrap variance estimator.  Default is 1000.}
}
\value{
A list of output containing:
\itemize{
\item{pop_total:}{Estimate of population total}
\item{pop_mean:}{Estimate of the population mean (or proportion)}
\item{pop_total_var:}{Estimated variance of population total estimate}
\item{pop_mean_var:}{Estimated variance of population mean estimate}
\item{strat_ests:}{Table of total and mean estimates for each strata}
\item{weights.ps:}{Survey weights produced by PS}
}
}
\description{
Calculates a generalized regression estimator for a finite population mean/proportion or total based on sample data collected from a complex sampling design and auxiliary population data.
}
\references{
\insertRef{coc77}{mase} 

\insertRef{sar92}{mase}
}
