% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioEstimator.R
\name{ratioEstimator}
\alias{ratioEstimator}
\title{Compute a ratio estimator}
\usage{
ratioEstimator(y, xsample, xpop, datatype = "raw", pi = NULL, N = NULL,
  pi2 = NULL, var_est = FALSE, var_method = "HB", B = 1000)
}
\arguments{
\item{y}{vector of response variable of length n, or a matrix with 
dimension n * 1}

\item{xsample}{predictor vector of sampled data of length n}

\item{xpop}{Dataframe of population level auxiliary information.  Must come in the form of raw data, population totals or population means.}

\item{datatype}{Default to "raw", takes values "raw", "totals", and "means" for whether the user is providing the raw population values for x, the population totals for x, or the population means for x}

\item{pi}{Default to assume equal probability/simple random sampling, if unequal probability, requires vector of first-order inclusion probabilities of same length as y}

\item{N}{population size, if not provided estimate with the sum of the inverse inclusion probabilities}

\item{pi2}{= a n * n matrix of the joint inclusion probabilities.  Needed for the "HT" variance estimator}

\item{var_est}{Default to FALSE, logical for whether or not to compute estimate of variance}

\item{var_method}{Method to use when computing the variance estimate.  Options are "HB"= Hajek-Berger estimator, "HH" = Hansen-Hurwitz estimator, "HTSRS" = Horvitz-Thompson estimator under simple random sampling, "HT" = Horvitz-Thompson estimator, "bootstrapSRS" = bootstrap variance estimator under simple random sampling without replacement}

\item{B}{number of bootstrap samples if computing the bootstrap variance estimator.  Default is 1000.}
}
\value{
List of output containing:
\itemize{
\item{pop_total:}{Estimate of population total}
\item{pop_mean:}{Estimate of population mean}
\item{pop_total_var:}{ Estimated variance of population total estimate}
\item{pop_mean_var:}{ Estimated variance of population mean estimate}
}
}
\description{
Calculates a ratio estimator for a finite population mean/proportion or total based on sample data collected from a complex sampling design and auxiliary population data.
}
\references{
\insertRef{coc77}{mase} 

\insertRef{sar92}{mase}
}
