% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREG.R
\name{greg}
\alias{greg}
\title{Compute a generalized regression estimator}
\usage{
greg(y, xsample, xpop, pi = NULL, model = "linear", pi2 = NULL,
  var_est = FALSE, var_method = "HB", datatype = "raw", N = NULL,
  modelselect = FALSE, lambda = "lambda.min", B = 1000)
}
\arguments{
\item{y}{survey variable of interest. Should be a numeric vector or factor with two levels.}

\item{xsample}{dataframe of auxiliary data in the sample.  Number of rows should match length of y.}

\item{xpop}{dataframe of population level auxiliary information.  Must contain same names as xsample.  Can supply either unit level data (denoted as datatype = "raw") or aggregate data (denoted as either datatype = "totals" or datatype = "means")}

\item{pi}{first order inclusion probabilities of same length as y. If not supplied, simple random sampling is assumed.}

\item{model}{regression model to utilize. Options are "linear" or "logistic".}

\item{pi2}{a square matrix of the joint inclusion probabilities.  Needed for the "HT" variance estimator.}

\item{var_est}{a logical indicating whether or not to compute a variance estimate}

\item{var_method}{method to use when computing the variance estimate.  Options are "HB"= Hajek-Berger estimator, "HH" = Hansen-Hurwitz estimator, "HTSRS" = Horvitz-Thompson estimator under simple random sampling, "HT" = Horvitz-Thompson estimator, "bootstrapSRS" = bootstrap variance estimator under simple random sampling without replacement.}

\item{datatype}{form of population auxiliary data. Takes values "raw", "totals" or "means" for whether the user is providing population data at the unit level, aggregated to totals, or aggregated to means.}

\item{N}{population size, if not provided estimated to be the sum of the inverse inclusion probabilities}

\item{modelselect}{default to FALSE.  If TRUE, the Lasso is run and only the predictors with a non-zero Lasso coefficient are included in the model.}

\item{lambda}{Default to "lambda.min".  Only use if modelselect = TRUE. Takes values "lambda.min", which is the lambda value associated with the minimum cross validation error or "lambda.1se", which is the lamabda value which is one standard error away from the minimizing lambda and produces a sparser fit}

\item{B}{number of bootstrap samples if computing the bootstrap variance estimator.  Default is 1000.}
}
\value{
A list of output containing:
\itemize{
\item{pop_total:}{Estimate of population total}
\item{pop_mean:}{Estimate of the population mean (or proportion)}
\item{weights:}{Survey weights produced by GREG (linear model only)}
\item{pop_total_var:}{Estimated variance of population total estimate}
\item{pop_mean_var:}{Estimated variance of population mean estimate}
}
}
\description{
Calculates a generalized regression estimator for a finite population mean/proportion or total based on sample data collected from a complex sampling design and auxiliary population data.
}
\references{
\insertRef{cas76}{mase}

\insertRef{sar92}{mase}
}
