% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregElasticNet.R
\name{gregElasticNet}
\alias{gregElasticNet}
\title{Compute an elastic net regression estimator}
\usage{
gregElasticNet(y, xsample, xpop, pi = NULL, alpha = 1, model = "linear",
  pi2 = NULL, var_est = FALSE, var_method = "HB", datatype = "raw",
  N = NULL, lambda = "lambda.min", B = 1000, cvfolds = 10)
}
\arguments{
\item{y}{vector of response variable of length n, or a matrix with 
dimension n * 1}

\item{xsample}{vectors of observed data of length n}

\item{xpop}{Dataframe of population level auxiliary information.  Must come in the form of raw data, population totals or population means.}

\item{pi}{Default to assume equal probability/simple random sampling, if unequal probability, requires vector of first-order inclusion probabilities of same length as y}

\item{alpha}{mixing parameter for the lasso and ridge penalties in the elastic net.  When alpha = 1, uses only a lasso penalty.  When alpha=0, uses only a ridge penalty.}

\item{model}{regression model to utilize. User must choose 'linear' or 'logistic'}

\item{pi2}{= a n * n matrix of the joint inclusion probabilities.  Needed for the "HT" variance estimator}

\item{var_est}{Default to FALSE, logical for whether or not to compute estimate of variance}

\item{var_method}{Method to use when computing the variance estimate.  Options are "HB"= Hajek-Berger estimator, "HH" = Hansen-Hurwitz estimator, "HTSRS" = Horvitz-Thompson estimator under simple random sampling, "HT" = Horvitz-Thompson estimator, "bootstrapSRS" = bootstrap variance estimator under simple random sampling without replacement}

\item{datatype}{Default to "raw", takes values "raw", "totals", and "means" for whether the user is providing the raw population values for x, the population totals for x, or the population means for x}

\item{N}{population size, if not provided estimated to be the sum of the inverse inclusion probabilities}

\item{lambda}{Default to "lambda.min", takes values "lambda.min", which is the lambda value associated with the minimum cross validation error or "lambda.1se", which is the lamabda value which is one standard error away from the minimizing lambda and produces a sparser fit}

\item{B}{number of bootstrap samples if computing the bootstrap variance estimator.  Default is 1000.}

\item{cvfolds}{number of folds for cross-validation to pick lambda.}
}
\value{
A list of output containing:
\itemize{
\item{pop_total:}{Estimate of population total}
\item{coefficients:} {Elastic net coefficient estimates}
\item{pop_mean:}{Estimate of the population mean (or proportion)}
\item{pop_total_var:}{Estimated variance of population total estimate}
\item{pop_mean_var:}{Estimated variance of population mean estimate}
}
}
\description{
Calculates a lasso, ridge or elastic net generalized regression estimator for a finite population mean/proportion or total based on sample data collected from a complex sampling design and auxiliary population data.
}
\references{
\insertRef{mcc17}{mase}
}
