% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregTree.R
\name{gregTree}
\alias{gregTree}
\title{Compute a regression tree estimator}
\usage{
gregTree(y, xsample, xpop, pi = NULL, pi2 = NULL, var_est = FALSE,
  var_method = "HB", B = 1000, pval = 0.05, perm_reps = 500,
  bin_size = NULL)
}
\arguments{
\item{y}{vector or a matrix with one column of the sampled response variable}

\item{xsample}{matrix of auxiliary data in the sample.  Number of rows should match length of y.}

\item{xpop}{Dataframe of population level auxiliary information.}

\item{pi}{Default to assume equal probability/simple random sampling, if unequal probability, requires vector of first-order inclusion probabilities of same length as y}

\item{pi2}{a square matrix of the joint inclusion probabilities.  Needed for the "HT" variance estimator}

\item{var_est}{Default to FALSE, logical for whether or not to compute estimate of variance}

\item{var_method}{Method to use when computing the variance estimate.  Options are "HB"= Hajek-Berger estimator, "HH" = Hansen-Hurwitz estimator, "HTSRS" = Horvitz-Thompson estimator under simple random sampling, "HT" = Horvitz-Thompson estimator, "bootstrapSRS" = bootstrap variance estimator under simple random sampling without replacement}

\item{B}{number of bootstrap samples if computing the bootstrap variance estimator.  Default is 1000.}

\item{pval}{Designated pval level to reject null hypothesis in permutation test used to fit the regression tree, Default to 0.05}

\item{perm_reps}{integer specifying the number of permutations for each permutation test run to fit the regression tree, Default to 500}

\item{bin_size}{numeric minimum number of observations in each node}

\item{N}{population size, if not provided estimated to be the sum of the inverse inclusion probabilities}
}
\value{
A list of output containing:
\itemize{
\item{pop_total:}{Estimate of population total}
\item{pop_mean:}{Estimate of the population mean (or proportion)}
\item{weights:}{Survey weights produced by gregTree}
\item{pop_total_var:}{Estimated variance of population total estimate}
\item{pop_mean_var:}{Estimated variance of population mean estimate}
}
}
\description{
Calculates a regression tree estimator for a finite population mean/proportion or total based on sample data collected from a complex sampling design and auxiliary population data.
}
\references{
\insertRef{mcc17b}{mase}
}
