% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horvitzThompson.R
\name{horvitzThompson}
\alias{horvitzThompson}
\title{Compute the Horvitz-Thompson Estimator}
\usage{
horvitzThompson(y, pi = NULL, N = NULL, pi2 = NULL, var_est = FALSE,
  var_method = "HB", B = 1000)
}
\arguments{
\item{y}{A numeric vector or a matrix with one column of the sampled response variable.}

\item{pi}{Default assumes equal probability/simple random sampling. If unequal probability, requires vector of first-order inclusion probabilities of same length as y}

\item{N}{population size, if not provided estimate with the sum of the inverse inclusion probabilities}

\item{pi2}{= a square matrix of the joint inclusion probabilities.  Needed for the "HT" variance estimator}

\item{var_est}{Default is FALSE, logical indicating whether or not to compute a variance estimator}

\item{var_method}{Method to use when computing the variance estimator.  Options are "HB"= Hajek-Berger estimator, "HH" = Hansen-Hurwitz estimator, "HTSRS" = Horvitz-Thompson estimator under simple random sampling without replacement, "HT" = Horvitz-Thompson estimator, "bootstrapSRS" = bootstrap variance estimator under simple random sampling without replacement}

\item{B}{number of bootstrap samples if computing the bootstrap variance estimator.  Default is 1000.}
}
\value{
List of output containing:
\itemize{
\item{pop_total:}{Estimate of population total}
\item{pop_mean:}{Estimate of population mean}
\item{pop_total_var:}{ Estimated variance of population total estimate}
\item{pop_mean_var:}{ Estimated variance of population mean estimate}
}
}
\description{
Calculate the Horvitz-Thompson Estimator for a finite population mean/proportion or total based on sample data collected from a complex sampling design.
}
\references{
\insertRef{hor52}{mase}
}
